netwerk/test/mochitests/test_signed_web_packaged_app.html

<!DOCTYPE html>
<html>
<head>
  <title> Web packaged app </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

var Cc = SpecialPowers.Cc;
var Ci = SpecialPowers.Ci;
var Cu = SpecialPowers.Cu;
var Cr = SpecialPowers.Cr;

SpecialPowers.pushPrefEnv(
  { "set": [["network.http.enable-packaged-apps", true],
            ["network.http.packaged-apps-developer-mode", true],
            ["network.http.packaged-signed-apps-enabled", true],
            ["dom.ipc.processPriorityManager.testMode", true],
            ["dom.ipc.processPriorityManager.enabled", true],
            ["dom.ipc.tabs.disabled", false],
            ["dom.ipc.processCount", 3],
            ["dom.mozBrowserFramesEnabled", true]] },
  () => SpecialPowers.pushPermissions([
    { "type": "browser", "allow": 1, "context": document }
  ], function() {
    runTest();
  }));

SimpleTest.waitForExplicitFinish();

// Listen for and count process-created event. Since we are loading a
// signed content from a remote tab, there shouls be two processes created.
// One is for remote tab and one for the signed package.
var kProcessCreatedTopic = "process-priority-manager:TEST-ONLY:process-created";
var processCreatedCnt = 0;
SpecialPowers.addObserver(() => {
  processCreatedCnt++;
  if (processCreatedCnt == 1) {
    ok(true, "We have one more process to create.");
  } else if (processCreatedCnt == 2) {
    SimpleTest.finish();
  }
}, kProcessCreatedTopic, /* weak = */ false);

function runTest() {
  var iframe = document.createElement("iframe");
  iframe.setAttribute('mozbrowser', 'true');
  iframe.setAttribute('remote', 'true');
  iframe.setAttribute("src", "http://example.org:80");

  iframe.addEventListener("mozbrowserloadend", function loadend(e) {
    iframe.removeEventListener("mozbrowserloadend", loadend);
    ok(true, "Got mozbrowserloadend");
    iframe.setAttribute("src", "http://mochi.test:8888/tests/netwerk/test/mochitests/signed_web_packaged_app.sjs!//scripts/app.js");
  });

  document.body.appendChild(iframe);
}

</script>
</pre>
</body>
</html>
